Tests/LibWeb/Text/expected/scrollable-overflow-negative-start-padding.txt

client width: 208
scroll width: 208
scroll left: 0
